+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-02-16 18:42:13
+
+### Changed Files
+- `src/features/admin/components/lead-vendors/LeadPoliciesTable.tsx`
+
+### Statistics
+```
+ 1 file changed, 15 insertions(+), 6 deletions(-)
+```
+
 ## [Checkpoint] - 2026-02-16 18:23:40
 
 ### Changed Files